move pathc path function to components command and remove check_components_structure
