+++ nf_core/components/components_command.py
@@ -158,22 +158,22 @@ class ComponentCommand:
         except FileNotFoundError:
             log.debug(f"No lint config file found: {config_fn}")
 
-    def check_component_structure(self, component_name):
+    def check_modules_structure(self):
         """
-        Check that the structure of the modules/subworkflow directory in a pipeline is the correct one:
-            modules/nf-core/TOOL/SUBTOOL | subworkflows/nf-core/SUBWORKFLOW
+        Check that the structure of the modules directory in a pipeline is the correct one:
+            modules/nf-core/TOOL/SUBTOOL
 
         Prior to nf-core/tools release 2.6 the directory structure had an additional level of nesting:
             modules/nf-core/modules/TOOL/SUBTOOL
         """
         if self.repo_type == "pipeline":
             wrong_location_modules = []
-            for directory, _, files in os.walk(Path(self.dir, component_name)):
+            for directory, _, files in os.walk(Path(self.dir, "modules")):
                 if "main.nf" in files:
-                    module_path = Path(directory).relative_to(Path(self.dir, component_name))
+                    module_path = Path(directory).relative_to(Path(self.dir, "modules"))
                     parts = module_path.parts
                     # Check that there are modules installed directly under the 'modules' directory
-                    if parts[1] == component_name:
+                    if parts[1] == "modules":
                         wrong_location_modules.append(module_path)
             # If there are modules installed in the wrong location
             if len(wrong_location_modules) > 0:
@@ -185,12 +185,43 @@ class ComponentCommand:
                 )
                 # Move wrong modules to the right directory
                 for module in wrong_location_modules:
-                    modules_dir = Path(component_name).resolve()
+                    modules_dir = Path("modules").resolve()
                     correct_dir = Path(modules_dir, self.modules_repo.repo_path, Path(*module.parts[2:]))
                     wrong_dir = Path(modules_dir, module)
                     shutil.move(wrong_dir, correct_dir)
                     log.info(f"Moved {wrong_dir} to {correct_dir}.")
-                shutil.rmtree(Path(self.dir, component_name, self.modules_repo.repo_path, component_name))
+                shutil.rmtree(Path(self.dir, "modules", self.modules_repo.repo_path, "modules"))
                 # Regenerate modules.json file
                 modules_json = ModulesJson(self.dir)
                 modules_json.check_up_to_date()
+
+    def check_patch_paths(self, patch_path, module_name):
+        """
+        Check that paths in patch files are updated to the new modules path
+        """
+        if patch_path.exists():
+            log.info(f"Modules {module_name} contains a patch file.")
+            rewrite = False
+            with open(patch_path, "r") as fh:
+                lines = fh.readlines()
+                for index, line in enumerate(lines):
+                    # Check if there are old paths in the patch file and replace
+                    if f"modules/{self.modules_repo.repo_path}/modules/{module_name}/" in line:
+                        rewrite = True
+                        lines[index] = line.replace(
+                            f"modules/{self.modules_repo.repo_path}/modules/{module_name}/",
+                            f"modules/{self.modules_repo.repo_path}/{module_name}/",
+                        )
+            if rewrite:
+                log.info(f"Updating paths in {patch_path}")
+                with open(patch_path, "w") as fh:
+                    for line in lines:
+                        fh.write(line)
+                # Update path in modules.json if the file is in the correct format
+                modules_json = ModulesJson(self.dir)
+                modules_json.load()
+                if modules_json.has_git_url_and_modules():
+                    modules_json.modules_json["repos"][self.modules_repo.remote_url]["modules"][
+                        self.modules_repo.repo_path
+                    ][module_name]["patch"] = str(patch_path.relative_to(Path(self.dir).resolve()))
+                modules_json.dump()
